requirements.txt
langchain
langgraph
chromadb
playwright
google-generativeai
python-dotenv
